Price Lib: Fix `i32` cannot be nullable

+++ subgraphs/_reference_/src/prices/common/constants.ts
@@ -311,7 +311,7 @@ CHAIN_LINK_CONTRACT_ADDRESS.set(
 ///////////////////////////////// HELPERS /////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////
 
-export const USDC_DECIMALS_MAP = new TypedMap<string, i32>();
+export const USDC_DECIMALS_MAP = new TypedMap<string, BigInt>();
 USDC_DECIMALS_MAP.set(MAINNET.NETWORK_STRING, MAINNET.USDC_DECIMALS);
 USDC_DECIMALS_MAP.set(BSC.NETWORK_STRING, BSC.USDC_DECIMALS);
 USDC_DECIMALS_MAP.set(FANTOM.NETWORK_STRING, FANTOM.USDC_DECIMALS);
